feat(build): Add a docker container that builds the opentrons conf (#2)
The opentrons-build.sh utility will build a docker container from the local
buildfile to run make in, and then run make in it. Arguments can be forwarded to
the docker run invocation; the final argument is passed as an argument to the
executable form of docker run (e.g. passed to make inside the container), so if
you just want docker arguments you should append "" to the opentrons-build.sh
invocation.

In addition, the default site for some packages, snapshot.debian.org, blacklists
an entire /9 of ec2 related IPs:
https://www.mail-archive.com/debian-bugs-dist@lists.debian.org/msg1638894.html

The change to fakeroot.mk switches the site to using deb.debian.org, which might
remove the specified version of the package at some time in the future but at
least it works now.