feat(config): Add opentrons ot2 board dir and defconfig
We now have a board dir under board/opentrons/ot2 and an ot2_defconfig to
comply with the normal buildroot method of doing things.

There are also some config tweaks in the defconfig, like

- Using linaro instead of building our own gcc every time
- Some tweaks to the partition layout from the general raspi config (definitely
  not done yet though)
- Use of systemd rather than buildroot init
- Full bash shell

There are also some changes to the docker infrastructure to quiet down the
output and make sure the structure is correct.

Closes #2863