chore(build): Split rootfs into root and var partitions (#4)
Since we want to update entire partitions while keeping some user data, we have to split the filesystem apart into multiple locations.

One partition (the "home" partition) is mounted at /var. /var/home is bind mounted to /home and /var/data is bind mounted to /data. This is in the last partition, which will be autoexpanded, and will never change - this is the persistent fs.

The root filesystem, which has everything that isn't in /var, /home, or /data, is determined by the kernel command line and is one of either of the root partitions. These are the ones that will be overwritten in updates, and they are mounted readonly.

Right now, this is all accomplished by

- Adding a /var image to the genimage config
- Removing /var from the staging fakeroot environment it's put into the root filesystem image
- Modifying /etc/fstab in post-build.sh

This all may change as things get more complex.

Closes https://github.com/Opentrons/opentrons/issues/2866