feat(system): Add dropbear
Dropbear serves ssh for us. It is configured to only allow logins with ssh keys.
Right now, a single ssh public key is put in authorized_keys during build. In
the future, we'll remove this for release builds and implement ways to add and
remove keys from local connections only.