feat(system): fall back to linklocal if dhcp fails on eth0
The easiest way to do this is to configure two connections in NetworkManager.
The first, with a higher priority, is for dhcp and gives up if it fails; the
second, with a normal priority, is for link-local and tries forever (until the
carrier goes away). This handles both kinds of connection.