feat(system): Add python dependencies for jupyter
There are a whole bunch. There also needed to be some packages modified to get
more recent versions, and a lot of interpolation of dependencies and transitive
dependencies to get buildroot to build everything correctly since jupyter stuff
has a nasty habit of including itself in its setup scripts, which makes us have
to install host dependencies to make sure we can build.

This also checks in some changes to scanpypi that make it a little more robust.