feat(system): Build, configure, and run jupyter notebook server
This enables jupyter notebook and adds a unit file to get it running on startup.
The unit file is a little more complex than in docker because we're actually
running jupyter in a service account (i.e: no default shell, no home directory)
rather than just as root. So we leverage systemd capabilities to create standard
directories.