feat(system): add system packages (#9)
* feat(system): Add various little system capabilities

- avrdude for module updates
- ffmpeg and mpg123 for camera and audio
- a cute motd

* feat(system): Add dropbear

Dropbear serves ssh for us. It is configured to only allow logins with ssh keys.
Right now, a single ssh public key is put in authorized_keys during build. In
the future, we'll remove this for release builds and implement ways to add and
remove keys from local connections only.

* feat(system): Add avahi and advertise on mdns

This implements our mdns advertising using static configurations. We'll
advertise as opentrons-(hostname).

* feat(system): add nginx

Adds nginx and the nginx.conf from the monorepo (but without specifying the pid,
since that's handled by the default unit file).

* feat(system): fall back to linklocal if dhcp fails on eth0

The easiest way to do this is to configure two connections in NetworkManager.
The first, with a higher priority, is for dhcp and gives up if it fails; the
second, with a normal priority, is for link-local and tries forever (until the
carrier goes away). This handles both kinds of connection.

* feat(system): Run boot.d scripts on startup

Executable scripts in /data/boot.d will be run after filesystems are mounted.

* feat(system): Add opentrons.target as default and fixup systemd

It always feels a little weird just using multi-user.target for things that
aren't really strictly necessary for multi-user work; and it was definitely
weird to be using graphical.target since we don't have graphics; so we have
opentrons.target as our new default systemd target, and added the thing that
runs boot.d as a want.

Also, fixup a couple other annoying systemd things we hadn't previously caught:
- systemd ExecStart parsing doesn't work like a shell unless you specifically
say to run it in a shell
- as much as you might want it to, putting a WantedBy in a unit only does
something if you're running systemctl enable on it, and we're not, so we need
som explicit symlinks
- it turns out echo and cat are different and we are now successfully setting up
authorized keys
- home is /root not /home

* feat(system): Add smoothie programmer

This is available on github at https://github.com/capiman/lpc21isp but I don't want to add it as a package because the checked in binary already works.

Closes https://github.com/Opentrons/opentrons/issues/2873