feat(system): Set up gpio at boot
This adds the fairly ugly hack script ot-blinkenlights. Because the opentrons
module can't be imported cleanly (it reads configs, might do database migration,
etc) we just find it using importlib.find_loader, put the drivers subdirectory
in the pythonpath, and do the setupgpio tasks. This avoids the situation in
balena where the config migration and db migration actually happen when you run
setup_gpio.

This also blinks the button while you're loading. It runs as early as possible
and should blink during the long boot time while the system expands varfs if
necessary.