feat(system): send logs to datadog with settable level (#22)
* feat(system): Upload logs to datadog using syslog-ng

All the logs previously configured to syslog-ng now go to datadog. The datadog
creds are pulled at build time from either AWS or the environment. If those
creds are incorrect or not present, everything on the robot still works fine but
you just don't get the aggregation

* feat(system): Get the min log level for syslog from /var

Editing the /var/lib/syslog-ng/minlevel to enter a valid syslog level will
change the minimum level of logging sent upstream to datadog.