chore(build): Fuse migration and update files
Having separate migration and update files means users have to know the
difference when picking one to upload to a robot. The only difference between
them is the presence or absence of a boot partition and hash, which are both
very small when zipped.

Remove the concept of a migration image and add the boot partition
unconditionally to the system update file.

Closes https://github.com/Opentrons/opentrons/issues/3559