fix(system): prevent camera from becoming alsa default device
Sometimes when the device boots, the usb camera's audio interface will get
enumerated before the bcm sound hardware from the pi. It's also only a mic. That
means that when you try to play sound, alsa can't take the device (since it's
not an output driver) and fail. We could configure alsa to use index 1 by
default, but maybe the initialization will change or the camera won't be plugged
in. This change just tells alsa to never use a usb audio device as the default,
which should leave the platform hardware the only thing to use.