fix(system): handle spaces in notebook names in jupyter migration
The previous notebook migration script fell into the classic trap of parsing the
output of ls, and splitting the names of files into words if there were spaces.
The usual solution is to rely on the shell and do `for file in
/some/dir/glob/*`, but that's a bashism and we use ash. The best way to handle
this is just to not use a shell script, so this is a python script that does the
same thing and as a bonus handles spaces in filenames and is much (much) more
readable.