board/opentrons/ot2: always sign builds if key is available
We have only been signing builds if the build is a release (so, not most edge
builds). However, it's hard to go from a signed build to an unsigned build - it
involves editing files on the filesystem - and signing the builds doesn't
actually take that long, so let's just do it whenever the signing key is
available.