Relax consistency guarantees for bind-mounted Docker volumes.
On macOS, this improves the performance of the build from "literally unusuable" to merely "painfully slow."

As a result of this change, bad things may happen if you try to poke around in the build directories while the build is running.  But even with strict consistency, that would have been a bad idea for other reasons.

This shouldn't affect anything on platforms other than macOS.