opentrons: Fix build scripts for building locally on macOS (#64)
- When SIGNING_KEY is not provided, don't create a .signing-key file.
This fixes the build breaking when SIGNING_KEY is unset.  post-build.sh would think that a signing key was provided because the file existed, even though it was empty.
- Use #!/bin/bash instead of #!/bin/sh in post-build.sh.
This script does use Bash features - the `[[ ... ]]` syntax, at least.

This did cause an error, but, worryingly, the script would continue execution despite the -e option being set.  Apparently this is normal, documented Bash behavior for an unrecognized command following an if-statement.  :|
- Add .DS_Store to .gitignore, for macOS.