chore: Pin Docker to Debian 10 to work around out-of-date Buildroot (#132)
We were seeing build failures in CI like:

> No rule to make target '/buildroot/output/.br-external.mk`

This appears to have been caused by:

* Debian having recently released a new stable version, Debian 11.
* Our Docker image, debian:stable-slim, having auto-updated to track Debian 11.
* Debian 11 having updated GNU Make from v4.2.1 to v4.3.
* Our Buildroot fork missing several upstream commits to fix compatibility with GNU Make v4.3. (Try grepping upstream’s CHANGES file for "4.3".)

As a quick fix, this pins our docker image to Debian 10 so we get Make v4.2.1 back. Unfortunately, Debian doesn't really support installing an old version of a single package, so it's all-or-nothing. See https://unix.stackexchange.com/a/544434/7421.