configs/ot2_defconfig: config for python3.10
This has a couple different config elements than 3.7; explicitly
enabling openssl lets us compile things like the ssl and hash modules.
Since we have to use openssl anyway (and openssl 1.1.1 isn't as much of
a horrorshow as it used to be) dump libressl.

Some overlays that were into 3.7 now also need to be in 3.10.