package/python-pandas: bump to 1.4.1
Python 3.10 contains some recent PEPs that are intended to make the C
API a little more friendly to having multiple instances in the same
process; these took the form of making the inc/decref c macros not
return lvalues anymore, which broke some extensions like pandas. Update
pandas to a recent enough version that it can handle the change.