build: add sdk generation
Buildroot SDKs are relocatable sysroots containing host tooling for
cross compiling. With the BR2_PACKAGE_HOST_ENVIRONMENT_SETUP flag, it
includes a script that will overwrite standard environment variables in
a shell with those from the SDK. This can be used to provide the same
cross-compiler environment that Buildroot uses outside buildroot itself,
for application development.

Closes RCORE-206