board/opentrons/ot2/config.txt: restore ttyAMA0 to uart0 so the robot-server can talk to the smoothie board. (#177)
* board/opentrons/ot2/config.txt: restore ttyAMA0 to uart0 so the robot-server can talk to the smoothie board
we are using uart0 (GPIO pins 14, 15) to talk to the smoothie board, since usually bluetooth is mapped to uart0 we need the miniuart-bt overlay in the config to switch ttyAMA0 back to uart0. Unfortunetely the miniuart-bt overlay does not work in this kernel version and causes a boot failure, without it we cant communicate with the smoothie board. So instead of using miniuart-bt lets just remmap uart0 and uart1 using dtoverlay=uart0 to manually set the pins it should operate on.