fix: audio modules, alsa config for sound output
Before the update, I think we had a pretty simple audio setup and after
the update we were missing the driver modules for audio on our soc.
Bringing the driver modules back is pretty easy, but we now also need a
bunch of extra alsa configuration to make sure that the volume is at
100% because at 50% it's completely inaudible.

We can use amixer (NOT alsamixer, that's a TUI interactive mixer) to set
the volume high, and alsactl to store it, but we also need to make sure
we have somewhere for alsactl's state file, and then bam we've got
persistent volume across reboots.