python-numpy: inline legacy setuptools build
Unfortunately, numpy really needs to build with setuptools and it really
needs to build with old setuptools with specific commands, and this I
think is the least-blast-area way to accomplish that. Better would be
updating numpy, but that is also fraught for other reasons (not least
that we'd have to do it in openembedded also).