Revert "package/opentrons-temperature-module-firmware: use correct version link to v2.1.1 tempdeck binary. (#252)"
This reverts commit c5b8713dda53c8816b9378edb3208450b905b2b8.

The correct url is definitely the magdeck@v2.0.1 one; magdeck@v2.1.1
does not exist. It shouldn't be magdeck at all, but is because the
modules build pulled the wrong tag. This change was originally made I
think because buildroot isn't properly invalidating the download for
some reason and reusing the old module firmware file.