package/python-ipykernel: remove host jupyter
ipykernel needs the rest of the jupyter stack to be present when it
builds because it dynamically creates a json api spec for jupyter at
build time. This is pretty reasonable when not cross-compiling (and when
wanting to support an arbitrary version of jupyter-client) but needing
the jupyter stack for the host requires adding host build support for a
truly tremendous number of packages, and we only need to support exactly
one jupyter version, so just inline it.

We'll need to regenerate 0002-vendor-jupyter-api-for-6.1.12.patch
whenever we update the version of jupyter we install.