add and use a mechanism to defer py installs
This lets us put the biggest packages on the rootfs as zips. We can then
add a new system install directory on the varfs to our pythonpath and
have a systemd unit that installs the zips over there on boot before we
start the rest of the software. This should let us fix the issues with
third party packages like jupyter and pandas running us out of space.